+++ frontend/src/App.js
@@ -20,7 +20,7 @@ import {
   Admin,
   AdminRoles,
   AdminUsers,
-  AdminUsersEdit
+  AdminUserEdit
 } from './containers';
 
 import { store } from './store'
@@ -148,7 +148,7 @@ class App extends Component {
             <Route exact path="/admin" component={Admin} />
             <Route exact path="/admin/roles" component={AdminRoles} />
             <Route exact path="/admin/users" component={AdminUsers} />
-            <Route exact path="/admin/users/:id" component={AdminUsersEdit} />
+            <Route exact path="/admin/users/:id" component={AdminUserEdit} />
           </div>
         </Router >
       </Provider>